add spring-boot-actuator and adjust xml mapper file in same directory or failed  ,test ok

+++ src/main/java/com/example/demo/service/impl/UserService.java
@@ -17,7 +17,7 @@ public class UserService implements IUser {
 
     public User getInfo()
     {
-        Integer uid = new Integer(1);
+        Integer uid = new Integer(2);
         return  userMapper.selectByPrimaryKey(uid);
 
     }
